Checkpoint refactoring

+++ src/perft.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-#include "../include/chess.h"
+#include "../include/engine.h"
 #include "../include/prettyprint.h"
 
 /* Runs perft test for a given board and depth */
@@ -27,7 +27,7 @@ uint64_t perft(board_t* board, int depth) {
 }
 
 /* Runs perft divide test for a given board and depth */
-int perft_divide(board_t* board, int depth) {
+uint64_t perft_divide(board_t* board, int depth) {
     maxpq_t movelst;
     initialize_maxpq(&movelst);
     generate_moves(board, &movelst);
